Better countdown
Changed the JS/JQuery used for countdowm timers (from Lopadz's JQuery+MomentJS to Pablog1's). It's so much better now, the code is so much shorter. The downside is there's only 1 offset but it's easy to adjust the time manually due to only having 1 hour difference. Also updated how the timer and carousel look.